refactor createHashlessEntrypoints
